Move `With` into `modifiers` module, document `modifiers` module.
